MICNO-730: move all querying functions to adapter
