+++ tests/unit/adapter/test_job_service.py
@@ -3,7 +3,6 @@ import pytest
 from requests_mock import Mocker as RequestsMocker
 
 from job_executor.adapter import job_service
-from job_executor.app import query_for_jobs
 from job_executor.model.job import Job, JobParameters, UserInfo
 from job_executor.exception import HttpResponseError
 
@@ -147,7 +146,9 @@ def test_get_maintenance_status_error(requests_mock: RequestsMocker):
 def test_query_for_jobs(
     is_paused, expected_result, requests_mock, monkeypatch
 ):
-    monkeypatch.setattr("job_executor.app.is_system_paused", lambda: is_paused)
+    monkeypatch.setattr(
+        "job_executor.adapter.job_service.is_system_paused", lambda: is_paused
+    )
 
     # Always return built jobs even if system is paused
     # If system is paused, return empty list for queued and queued_manager jobs
@@ -163,5 +164,5 @@ def test_query_for_jobs(
         "job_executor.adapter.job_service.get_jobs", mock_get_jobs
     )
 
-    result = query_for_jobs()
+    result = job_service.query_for_jobs()
     assert result == expected_result
